requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.2
django-rest-swagger==0.3.2
django-toolbelt==0.0.1
djangorestframework==3.1.3
gunicorn==19.3.0
psycopg2==2.6.1
PyYAML==3.11
static3==0.6.1
wheel==0.24.0
